Separated the encoder and decoder
