+++ lib/src/summarizer_helper_methods.dart
@@ -2,7 +2,9 @@ import 'dart:typed_data';
 
 import 'package:flutter_local_summarizer/src/common_functions.dart';
 import 'package:flutter_local_summarizer/src/external_links.dart';
-import 'package:flutter_local_summarizer/src/model.dart';
+import 'package:flutter_local_summarizer/src/model/decoder.dart';
+import 'package:flutter_local_summarizer/src/model/encoder.dart';
+import 'package:flutter_local_summarizer/src/model/model.dart';
 import 'package:flutter_local_summarizer/src/tokenizer.dart';
 import 'package:onnxruntime/onnxruntime.dart';
 
@@ -42,11 +44,6 @@ class SummarizerHelperMethods {
     return summaryOutput;
   }
 
-  Future<OrtSession> _loadSession(Uint8List encoderList) async {
-    final sessionOptions = OrtSessionOptions();
-    return OrtSession.fromBuffer(encoderList, sessionOptions);
-  }
-
   Future<String?> _summary(
     String text, {
     int maxSummaryLength = 70,
@@ -71,10 +68,11 @@ class SummarizerHelperMethods {
     final OrtRunOptions runOptions = OrtRunOptions();
 
     printInDebug('Start');
-    final List<List<List<double>>>? outputs = await _generatEncode(
+    final List<List<List<double>>>? outputs = await Encoder.generateEncode(
       attentionMaskOrt: attentionMaskOrt,
       inputOrt: inputOrt,
       runOptions: runOptions,
+      model: encoderModel,
     );
 
     if (outputs == null) {
@@ -87,7 +85,7 @@ class SummarizerHelperMethods {
         OrtValueTensor.createTensorWithDataList(floatOutputs);
 
     // printInDebug(outputs);
-    List<int>? decodeInts = await generateDecode(
+    List<int>? decodeInts = await Decoder.generateDecode(
       attentionMaskOrt: attentionMaskOrt,
       inputOrt: inputOrt,
       runOptions: runOptions,
@@ -95,6 +93,7 @@ class SummarizerHelperMethods {
       maxSummaryLength: maxSummaryLength,
       eosTokenId: tokenizer.getEosTokenId(),
       progress: progress,
+      model: decoderModel,
     );
 
     if (decodeInts == null) {
@@ -142,134 +141,6 @@ class SummarizerHelperMethods {
     return inputList.map((sequence) {
       return sequence.map((token) => token != 0 ? 1 : 0).toList();
     }).toList();
-  }
-
-  Future<List<List<List<double>>>?> _generatEncode({
-    required OrtValueTensor inputOrt,
-    required OrtValueTensor attentionMaskOrt,
-    required OrtRunOptions runOptions,
-  }) async {
-    List<OrtValue?>? outputs;
-
-    final inputs = {
-      'input_ids': inputOrt,
-      'attention_mask': attentionMaskOrt,
-    };
-
-    printInDebug('Start generatEncode');
-    final OrtSession session = await _loadSession(encoderModel.biteList);
-    outputs = await session.runAsync(runOptions, inputs);
-    printInDebug('Done generatEncode');
-
-    if (outputs == null || outputs.isEmpty) {
-      return null;
-    }
-
-    final OrtValue? output0 = outputs[0];
-    if (output0 == null) {
-      return null;
-    }
-    final List<List<List<double>>> output0Value =
-        output0.value! as List<List<List<double>>>;
-
-    for (final element in outputs) {
-      element?.release();
-    }
-    session.release();
-
-    return output0Value;
-  }
-
-  Future<List<int>?> generateDecode({
-    required OrtValueTensor inputOrt,
-    required OrtValueTensor attentionMaskOrt,
-    required OrtValueTensor encodeOutput,
-    required OrtRunOptions runOptions,
-    required int maxSummaryLength,
-    required int eosTokenId,
-    Function(int)? progress,
-  }) async {
-    final OrtSession session = await _loadSession(decoderModel.biteList);
-    final List<int> currentOutput = [];
-
-    // Start with the initial decoder input
-    final List<int> initialDecoderInput = [
-      (inputOrt.value as List<List<int>>).first[0],
-    ];
-    currentOutput.addAll(initialDecoderInput);
-
-    printInDebug('Start generateDecode');
-
-    for (int i = 0; i < maxSummaryLength; i++) {
-      progress?.call((((i + 1) / maxSummaryLength) * 100).toInt());
-
-      // Prepare inputs for the decoder
-      final inputs = {
-        'input_ids': OrtValueTensor.createTensorWithDataList([currentOutput]),
-        'encoder_attention_mask': attentionMaskOrt,
-        'encoder_hidden_states': encodeOutput,
-      };
-
-      // Run the decoder
-      final List<OrtValue?>? outputs =
-          await session.runAsync(runOptions, inputs);
-
-      if (outputs == null || outputs.isEmpty) {
-        printInDebug('Decoder outputs are empty!');
-        break;
-      }
-
-      // Extract logits and find the next token
-      final OrtValue? output0 = outputs[0];
-      if (output0 == null) {
-        printInDebug('Decoder output[0] is null!');
-        break;
-      }
-      final List<List<List<double>>> output0Value =
-          output0.value! as List<List<List<double>>>;
-      final List<int> nextTokenIds = _npArgmax(output0Value[0]);
-      final int nextTokenId = nextTokenIds.last; // Get the last token ID
-
-      // Append the new token to the current output
-      currentOutput.add(nextTokenId);
-
-      // Release outputs to free resources
-      for (final element in outputs) {
-        element?.release();
-      }
-
-      // Stop if the EOS token is generated
-      if (nextTokenId == eosTokenId) {
-        printInDebug('EOS token encountered. Stopping decoding.');
-        break;
-      }
-    }
-
-    printInDebug('Done generateDecode');
-    session.release();
-
-    return currentOutput;
-  }
-
-  /// Using axis -1
-  List<int> _npArgmax(List<List<double>> logits) {
-    final List<int> maxIndices = [];
-
-    for (final List<double> innerList in logits) {
-      int maxIndex = 0;
-      double maxValue = innerList[0];
-
-      for (int i = 1; i < innerList.length; i++) {
-        if (innerList[i] > maxValue) {
-          maxValue = innerList[i];
-          maxIndex = i;
-        }
-      }
-
-      maxIndices.add(maxIndex);
-    }
-
-    return maxIndices;
   }
 
   // String _preprocessText(String text) {
